pyproject.toml
[project]
name = "laughing-invention"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-playwright>=0.7.0",
]
